// src/typings/imagem.d.ts
import { Variavel } from './variavel';

type Image = {
    approval: boolean;
    approved: boolean;
    category: string;
    closedFormat: string;
    editable: boolean;
    field: Variavel[];
    finalDetails: string;
    height: string;
    idImage: string;
    name: string;
    openFormat: string;
    requestDate: string;
    requester: string;
    s3Url: string;
    width: string;
}